{{cookiecutter.repo_name}}/{{cookiecutter.project_name}}/config/urls.py
from django.conf import settings
from django.conf.urls import url, include
from django.conf.urls.static import static
from django.contrib import admin

urlpatterns = [
    url(r'^admin/', admin.site.urls),
    url(r'^api-auth/', include('rest_framework.urls', namespace='rest_framework')),
    url(r'^api/', include('{{cookiecutter.project_name}}.config.routers.api_urls', namespace='api')),
    url(r'^', include('{{cookiecutter.project_name}}.config.routers.web_urls'))
] + static(settings.MEDIA_URL, document_root=settings.MEDIA_ROOT)


if settings.DEBUG:
    import debug_toolbar
    urlpatterns += [
        url(r'^__debug__/', include(debug_toolbar.urls))
    ]
