{{cookiecutter.repo_name}}/requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --output-file requirements.txt requirements.in
#
django-cors-headers==2.1.0
django-debug-toolbar==1.8
django-environ==0.4.3
django-extensions==1.8.1
django==1.11.4
djangorestframework-jwt==1.11.0
djangorestframework==3.6.3
gevent==1.2.2
greenlet==0.4.12          # via gevent
gunicorn==19.7.1
olefile==0.44             # via pillow
pillow==4.2.1
psycopg2==2.7.3
pyjwt==1.5.2              # via djangorestframework-jwt
pytz==2017.2              # via django
six==1.10.0               # via django-environ, django-extensions
sqlparse==0.2.3           # via django-debug-toolbar
